Move game3.h, game3_shared.h to inc/common/
