+++ src/server/game3_proxy/game3_pmove.c
@@ -17,8 +17,8 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 */
 
 #include "shared/shared.h"
+#include "shared/game3_shared.h"
 #include "common/pmove.h"
-#include "common/game3_shared.h"
 #include "game3_pmove.h"
 
 #define STEPSIZE    18
